% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v4C.R
\name{getHicSparse}
\alias{getHicSparse}
\title{getHic}
\usage{
getHicSparse(hic, gr, norm = "KR", type = "BP", res = 10000, mc.cores = 1, ...)
}
\arguments{
\item{hic}{path to .hic file}

\item{gr}{granges to query}

\item{norm}{string specifying normalization to apply ("KR" (default), "NONE", VC", "VC_SQRT")}

\item{type}{"BP" vs "FRAG"}

\item{res}{resolution to query (default 10kb)}

\item{mc.cores}{parallelization to apply for query (useful when length(gr)>2)}
}
\value{
interactionSet
}
\description{
queries .hic object via straw API https://github.com/theaidenlab/straw/tree/master/R
to return sparse upper triangle matrix
}
\author{
Ashley S Doane
if (n>1)
{
  pairs = t(combn(n, 2)) ##pairs
  grs.pairs = paste(grs[pairs[,1]], grs[pairs[,2]])   ## combinations
  #' Tuesday, Nov 07, 2017 04:43:52 PM - Julie: adding as.character()
  chr1.pairs = as.character(seqnames(gr)[pairs[,1]])
  chr2.pairs = as.character(seqnames(gr)[pairs[,2]])
}
}
\keyword{straw}
