% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v4C.R
\name{getmultiHic}
\alias{getmultiHic}
\title{getHic}
\usage{
getmultiHic(
  hic,
  grs,
  norm = "VC_SQRT",
  type = "BP",
  res = 10000,
  mc.cores = 1,
  ...
)
}
\arguments{
\item{hic}{path to .hic file}

\item{norm}{string specifying normalization to apply ("KR" (default), "NONE", VC", "VC_SQRT")}

\item{type}{"BP" vs "FRAG"}

\item{res}{resolution to query (default 10kb)}

\item{mc.cores}{parallelization to apply for query (useful when length(gr)>2)}

\item{gr}{granges to query}
}
\value{
interactionSet
}
\description{
queries .hic object via straw API https://github.com/theaidenlab/straw/tree/master/R
to return sparse upper triangle matrix
}
\author{
Ashley S Doane
}
\keyword{straw}
