% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v4C.R
\name{getViewPt}
\alias{getViewPt}
\title{getViewPt}
\usage{
getViewPt(
  hic,
  gr,
  norm = "KR",
  type = "BP",
  res = 5000,
  fwd = 20,
  rev = 20,
  mc.cores = 1,
  ...
)
}
\arguments{
\item{hic}{path to .hic file}

\item{gr}{granges to query representing viewpoint}

\item{norm}{string specifying normalization to apply ("KR" (default), "NONE", VC", "VC_SQRT")}

\item{type}{"BP" vs "FRAG"}

\item{res}{resolution to query (default 10kb)}

\item{fwd}{number of bins downstream of viewpoint}

\item{rev}{number of bins upstream of viewpoint}

\item{mc.cores}{parallelization to apply for query (useful when length(gr)>2)}
}
\value{
interactionSet
}
\description{
queries .hic object via straw API https://github.com/theaidenlab/straw/tree/master/R
to return virtual 4C data
}
\author{
Ashley S Doane
}
\keyword{straw}
