% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v4C.R
\name{gi.straw}
\alias{gi.straw}
\title{gi.straw}
\usage{
gi.straw(hic, gr, norm = "KR", type = "BP", res = 10000, mc.cores = 1, ...)
}
\arguments{
\item{hic}{path to .hic file}

\item{gr}{granges to query}

\item{norm}{string specifying normalization to apply ("KR" (default), "NONE", VC", "VC_SQRT")}

\item{type}{"BP" vs "FRAG"}

\item{res}{resolution to query (default 10kb)}

\item{mc.cores}{parallelization to apply for query (useful when length(gr)>2)}
}
\value{
interactionSet
}
\description{
queries .hic object via straw API https://github.com/theaidenlab/straw/tree/master/R
to return interactionSet
}
\author{
Ashley S Doane
Tuesday, Nov 07, 2017 04:43:52 PM - Julie: adding as.character()
}
\keyword{straw}
