% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v4C.R
\name{readDeLoop}
\alias{readDeLoop}
\alias{readCLoop}
\title{getHic}
\usage{
readDeLoop(
  deLoop = "/mnt/athena/projectshg38/datasets/hichip/cloop/H3K27acControl.deloop"
)

readCLoop(
  deLoop = "/mnt/athena/projectshg38/datasets/hichip/cloop/H3K27acControl.deloop"
)
}
\arguments{
\item{deLoop}{path to deLoop file}
}
\value{
GInteractions object

GInteractions object
}
\description{
reads cLoops differential loops
to return GInteractionSet

reads cLoops differential loops
to return GInteractionSet
}
\author{
Ashley S Doane

Ashley S Doane
}
\keyword{straw}
